modify connection pool
