+++ backend/pg_pool copy.js
@@ -1,0 +1,10 @@
+// const { Client } = require('pg')
+// const client = new Client({
+//     user: process.env.DB_USER,
+//     host: process.env.DB_HOST,
+//     database: process.env.DB_NAME,
+//     password: process.env.DB_PW,
+//     port: process.env.DB_PORT,
+// });
+
+// module.exports = client;
